api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
stripe==11.5.0
boto3==1.36.12
pytest==8.3.4
httpx==0.28.1
email-validator==2.2.0
